Move CreateSalesforceLead up a module
