+++ spec/features/newflow/educators/educator_signed_params_flow_spec.rb
@@ -4,7 +4,7 @@ module Newflow
   feature 'Educator signed params flow', js: true do
     before do
       turn_on_educator_feature_flag
-      allow_any_instance_of(Newflow::EducatorSignup::CreateSalesforceLead).to receive(:exec).and_return(true)
+      allow_any_instance_of(Newflow::CreateSalesforceLead).to receive(:exec).and_return(true)
     end
 
     background do
